XD-2644: Per module pligins
Introduce per-module plugin context

Remove non-binding plugin

reverse dependency

Remove spark

Remove filter, key not needed anymore

make dirt provided

Add packaging step to test modules

Extract test modules as their own project
