+++ extensions/spring-xd-extension-spark-streaming-examples/src/main/java/org/springframework/xd/spark/streaming/examples/java/FileLoggerModuleOptionsMetadata.java
