+++ extensions/spring-xd-extension-spark-streaming-examples/src/main/scala/org/springframework/xd/spark/streaming/examples/scala/FileLogger.scala
