+++ spring-xd-dirt/src/main/java/org/springframework/xd/dirt/plugins/NoOpPlugin.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2015 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *
+ */
+
+package org.springframework.xd.dirt.plugins;
+
+import org.springframework.xd.module.core.Module;
+
+/**
+ * A plugin that does nothing and does not trigger. Useful when overriding other plugins by name.
+ *
+ * @author Eric Bottard
+ */
+public class NoOpPlugin extends AbstractPlugin{
+	@Override
+	public boolean supports(Module module) {
+		return false;
+	}
+}
