+++ spring-xd-module/src/main/java/org/springframework/xd/module/core/AbstractModule.java
@@ -71,8 +71,4 @@ public abstract class AbstractModule implements Module {
 				ObjectUtils.getIdentityHexString(this) + "]";
 	}
 
-	@Override
-	public boolean shouldBind() {
-		return true;
-	}
 }
