+++ spring-xd-module/src/main/java/org/springframework/xd/module/core/Module.java
@@ -118,9 +118,4 @@ public interface Module extends Lifecycle {
 	 */
 	void destroy();
 
-	/**
-	 * Should the module require messagebus binding.
-	 */
-	boolean shouldBind();
-
 }
