+++ spring-xd-spark-streaming-tests/modules/processor/spark-scala-word-count/config/spark-scala-word-count.properties
@@ -1,0 +1,2 @@
+options_class=org.springframework.xd.spark.streaming.DefaultSparkStreamingModuleOptionsMetadata
+module.classpath=/lib/*.jar , ${xd.home}/spring-xd-spark-streaming/build/libs/spring-xd-spark-streaming-*.jar
