+++ spring-xd-spark-streaming-tests/modules/processor/spark-scala-word-count/config/spark-scala-word-count.xml
