+++ spring-xd-spark-streaming-tests/modules/processor/spark-word-count/config/spark-word-count.xml
