+++ spring-xd-spark-streaming-tests/modules/sink/file-logger-scala/config/file-logger-scala.properties
@@ -1,0 +1,2 @@
+options_class=org.springframework.xd.spark.streaming.examples.java.FileLoggerModuleOptionsMetadata
+module.classpath=/lib/*.jar , ${xd.home}/spring-xd-spark-streaming/build/libs/spring-xd-spark-streaming-*.jar
