+++ spring-xd-spark-streaming-tests/modules/sink/file-logger-scala/config/file-logger-scala.xml
