+++ spring-xd-spark-streaming-tests/modules/sink/file-logger/config/file-logger.xml
