--- spring-xd-spark-streaming-tests/src/test/resources/spring-xd/xd/modules/processor/spark-scala-word-count/config/spark-scala-word-count.properties
@@ -1,1 +1,0 @@
-options_class=org.springframework.xd.spark.streaming.DefaultSparkStreamingModuleOptionsMetadata
