--- spring-xd-spark-streaming-tests/src/test/resources/spring-xd/xd/modules/sink/file-logger/config/file-logger.properties
@@ -1,1 +1,0 @@
-options_class=org.springframework.xd.spark.streaming.examples.java.FileLoggerModuleOptionsMetadata
