+++ spring-xd-spark-streaming/src/main/java/org/springframework/xd/dirt/plugins/spark/streaming/MessageBusReceiver.java
