+++ spring-xd-spark-streaming/src/main/resources/META-INF/spring-xd/module-plugins/spark-streaming.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans"
+	   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	   xmlns:context="http://www.springframework.org/schema/context"
+	   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+	   http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd">
+
+	<context:annotation-config />
+
+	<bean id="streamPlugin" class="org.springframework.xd.dirt.plugins.spark.streaming.SparkStreamingPlugin"/>
+
+	<bean id="moduleTypeConversionPlugin" class="org.springframework.xd.dirt.plugins.NoOpPlugin" />
+
+</beans>
