--- spring-xd-spark-streaming/src/main/resources/META-INF/spring-xd/plugins/spark-streaming.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-	   xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	   xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd">
-
-	<bean class="org.springframework.xd.dirt.plugins.spark.streaming.SparkStreamingPlugin"/>
-
-</beans>
